Refactor chatgpt module to separate files
Split the large chatgpt/mod.rs monolith into focused modules:
- client.rs: GptClient implementation and API interaction logic
- message.rs: Message types and content handling
- role.rs: Role enum with serialization
- request.rs: ChatRequest structure
- response.rs: Response parsing and error handling

Updated module exports and fixed broken tests to work with new structure.
