+++ src/chatgpt/request.rs
@@ -1,0 +1,12 @@
+use serde::{Deserialize, Serialize};
+use crate::chatgpt::message::Message;
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ChatRequest {
+    pub model: String,
+    pub messages: Vec<Message>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub web_search_options: Option<serde_json::Value>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub max_tokens: Option<u32>,
+}
