+++ src/chatgpt/response.rs
@@ -1,0 +1,45 @@
+use serde::{Deserialize, Serialize};
+use serde_json::Result;
+use crate::chatgpt::message::Message;
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ChatResponse {
+    pub id: Option<String>,
+    pub object: Option<String>,
+    pub created: u64,
+    pub model: String,
+    pub usage: Usage,
+    pub choices: Vec<Choice>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ErrorResponse {
+    pub error: ErrorDetail,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct ErrorDetail {
+    pub message: String,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Usage {
+    pub prompt_tokens: u64,
+    pub completion_tokens: u64,
+    pub total_tokens: u64,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Choice {
+    pub message: Message,
+    pub finish_reason: String,
+    pub index: u64,
+}
+
+pub fn parse_response(json_str: &str) -> Result<ChatResponse> {
+    serde_json::from_str(json_str)
+}
+
+pub fn parse_error_response(json_str: &str) -> Result<ErrorResponse> {
+    serde_json::from_str(json_str)
+}
